+++ ui/CHANGES.md
@@ -1,5 +1,6 @@
 #### Version: 1.1.2 (LibGDX 1.9.3)
 - **Added**: `CursorManager`
+- **Added** `ScrollableTextArea` and `HighlightTextArea` with `Highlighter` API - Warning! Experimental, API and behaviour may change at any time!
 - **Misc**: Disabling Android Lint is no longer necessary
 
 #### Version: 1.1.1 (LibGDX 1.9.3)